requirements.txt
# ShowOnce Dependencies
# Install with: pip install -r requirements.txt

# Core
python-dotenv>=1.0.0          # Environment variable management
pydantic>=2.0.0               # Data validation and models
click>=8.1.0                  # CLI framework

# Image Processing
Pillow>=10.0.0                # Screenshot handling
imagehash>=4.3.0              # Image comparison (optional)

# AI Integration
anthropic>=0.18.0             # Claude API client

# Automation Frameworks
playwright>=1.40.0            # Browser automation (primary)
pyautogui>=0.9.54             # Desktop automation (fallback)

# Screen Capture
mss>=9.0.0                    # Fast cross-platform screenshots
pynput>=1.7.6                 # Keyboard/mouse listener for hotkeys

# Utilities
rich>=13.0.0                  # Beautiful terminal output
tqdm>=4.66.0                  # Progress bars

# Development
pytest>=7.4.0                 # Testing framework
black>=23.0.0                 # Code formatting
flake8>=6.1.0                 # Linting

# Optional: UI
streamlit>=1.29.0             # Web interface (Stage 5)
